pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt<4.0.0",
    "cloudinary>=1.42.2",
    "fastapi[standard]>=0.115.11",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]
